tools/ocr-gui/requirements.txt
customtkinter>=5.2.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
